+++ src/editors/TAudioScalerView.cpp
@@ -81,7 +81,7 @@ void TAudioScalerView::Init()
 
 	// Set up the array of rects for hitpoint detection.  We divide the area into
 	// a number of rects equal to the total number of ticks.
-	short tickWidth = fControlRect.Width();
+	short tickWidth = (short)(fControlRect.Width());
 	short spaceWidth = tickWidth / kTotalTicks;
 
 	for (short index = 0; index < kTotalTicks; index++) {
